Move native related files to /native.
